# documentazione/swagger-openapi.yaml
openapi: 3.0.1
info:
  title: OSP API
  description: Gestionale Occupazione Suolo Pubblico Bari
  version: '1.0'
servers:
  - url: http://localhost:8080/api/osp
    description: Generated server url
paths:
  /user-management/user:
    get:
      tags:
        - user-controller
      operationId: getUsers
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: username
          in: query
          required: false
          schema:
            maxLength: 30
            minLength: 3
            type: string
        - name: nome
          in: query
          required: false
          schema:
            maxLength: 50
            minLength: 0
            type: string
        - name: cognome
          in: query
          required: false
          schema:
            maxLength: 50
            minLength: 0
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUtenteDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    put:
      tags:
        - user-controller
      operationId: modificaDatiUtente
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UtenteEditRequest'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    post:
      tags:
        - user-controller
      operationId: inserisciUtente
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UtenteInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UtenteDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /user-management/user/password:
    put:
      tags:
        - user-controller
      operationId: cambiaPasswordUtente
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModificaPasswordRequest'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /user-management/user/{id}:
    get:
      tags:
        - user-controller
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UtenteDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    delete:
      tags:
        - user-controller
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /user-management/user/gruppo:
    get:
      tags:
        - user-controller
      operationId: getUsersGruppo
      parameters:
        - name: idGruppo
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UtenteDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /user-management/user/gruppo-municipio:
    get:
      tags:
        - user-controller
      operationId: getUsersGruppoMunicipio
      parameters:
        - name: idGruppo
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: idMunicipio
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UtenteDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
components:
  schemas:
    UtenteEditRequest:
      required:
        - codiceFiscale
        - email
        - enabled
        - id
        - idGruppo
        - username
      type: object
      properties:
        username:
          maxLength: 30
          minLength: 3
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
        idGruppo:
          type: integer
          format: int32
        idsMunicipi:
          type: array
          items:
            type: integer
            format: int32
        nome:
          type: string
        cognome:
          type: string
        sesso:
          type: string
          enum:
            - M
            - F
        dataDiNascita:
          type: string
          format: date
        luogoDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        provinciaDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        codiceFiscale:
          maxLength: 16
          minLength: 11
          type: string
        ragioneSociale:
          maxLength: 100
          minLength: 0
          type: string
        numTel:
          maxLength: 15
          minLength: 0
          type: string
        uoId:
          maxLength: 255
          minLength: 0
          type: string
        indirizzo:
          maxLength: 255
          minLength: 0
          type: string
        skipCheckConcessionario:
          pattern: ^$|^true$|^false$
          type: string
        id:
          type: integer
          format: int64
        enabled:
          type: boolean
    ModificaPasswordRequest:
      required:
        - oldPassword
        - password
        - username
      type: object
      properties:
        username:
          type: string
        oldPassword:
          maxLength: 40
          minLength: 0
          type: string
        password:
          maxLength: 40
          minLength: 6
          type: string
...
    UtenteDTO:
      required:
        - codiceFiscale
        - dateCreated
        - email
        - idGruppo
        - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        idGruppo:
          type: integer
          format: int32
        idsMunicipi:
          uniqueItems: true
          type: array
          items:
            type: integer
            format: int32
        username:
          maxLength: 30
          minLength: 3
          type: string
        nome:
          maxLength: 50
          minLength: 0
          type: string
        cognome:
          maxLength: 50
          minLength: 0
          type: string
        sesso:
          type: string
          enum:
            - M
            - F
        dataDiNascita:
          type: string
          format: date
        luogoDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        provinciaDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        codiceFiscale:
          maxLength: 16
          minLength: 11
          type: string
        ragioneSociale:
          maxLength: 50
          minLength: 0
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
        numTel:
          maxLength: 15
          minLength: 0
          type: string
        enabled:
          type: boolean
        dateCreated:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        flagEliminato:
          type: boolean
        dataEliminazione:
          type: string
          format: date-time
        uoId:
          maxLength: 255
          minLength: 0
          type: string
        indirizzo:
          maxLength: 255
          minLength: 0
          type: string
...
    UtenteInsertRequest:
      required:
        - codiceFiscale
        - email
        - idGruppo
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 30
          minLength: 3
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
        idGruppo:
          type: integer
          format: int32
        idsMunicipi:
          type: array
          items:
            type: integer
            format: int32
        nome:
          type: string
        cognome:
          type: string
        sesso:
          type: string
          enum:
            - M
            - F
        dataDiNascita:
          type: string
          format: date
        luogoDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        provinciaDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        codiceFiscale:
          maxLength: 16
          minLength: 11
          type: string
        ragioneSociale:
          maxLength: 100
          minLength: 0
          type: string
        numTel:
          maxLength: 15
          minLength: 0
          type: string
        uoId:
          maxLength: 255
          minLength: 0
          type: string
        indirizzo:
          maxLength: 255
          minLength: 0
          type: string
        skipCheckConcessionario:
          pattern: ^$|^true$|^false$
          type: string
        password:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageUtenteDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/UtenteDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    jwt:
      type: http
      in: header
      scheme: Bearer